% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairPlot.R
\name{PairPlot}
\alias{PairPlot}
\title{Design Projection Plot}
\usage{
PairPlot(Design, greys = TRUE)
}
\arguments{
\item{Design}{List object from MmSimplex, also works with a matrix input}

\item{greys}{Draws grey boxes to indicate the region of valid design points}
}
\value{
A ggplot object
}
\description{
Plots the projection of all pairs of inputs from a design from MmSimplex.  Can actually be used for any design on a Simplex.
}
\examples{
## Generate Design
D1 <- MmSimplex(3,75,10,cords = 1,randst = 1,phival = 50)

##Plot Design
PairPlot(D1)
}
